+++ lib/src/desktop/download_isolate.dart
@@ -37,10 +37,11 @@ Future<void> doDownloadTask(
   // Otherwise, it is a generated full path to the temp directory
   final tempFilePath = isResume && resumeData != null
       ? resumeData.tempFilepath
-      : p.join(
-          (await getTemporaryDirectory()).path,
-          'com.bbflight.background_downloader${Random().nextInt(1 << 32).toString()}',
-        );
+      : (task.options?.tempFilepath ??
+          p.join(
+            (await getTemporaryDirectory()).path,
+            'com.bbflight.background_downloader${Random().nextInt(1 << 32).toString()}',
+          ));
   final requiredStartByte =
       resumeData?.requiredStartByte ?? 0; // start for resume
   final eTag = resumeData?.eTag;